Struct loirc::Writer

pub struct Writer {
    // some fields omitted
}

Used to send messages to the IRC server.

This object is thread safe. You can clone it and send the clones to other threads. You can write from multiple threads without any issue. Internally, it uses Arc and Mutex.

Methods

impl Writer

fn disconnect(&self) -> Result<(), Error>

Drop the connection and trigger the reconnection process.

There might be a reconnection attempt, based on your settings. This should be used if you want the connection to be re-created. This is not the preferred way of shutting down the connection for good. Use close for this.

fn is_closed(&self) -> bool

Check if the connection was manually closed.

fn close(&self) -> Result<(), Error>

Close the connection and stop listening for messages.

There will not be any reconnection attempt. An error will be returned if the connection is already closed.

fn raw<S: AsRef<str>>(&self, data: S) -> Result<(), Error>

Send a raw string to the IRC server.

A new line will be not be added, so make sure that you include it. An error will be returned if the client is disconnected.